Function oak_client::interceptors::combine

pub fn combine<A: Interceptor, B: Interceptor>(
    interceptor_a: A, 
    interceptor_b: B
) -> CombinedInterceptor<A, B>

Combines the two provided interceptor, executing the first one, and if that succeeds then also the second one, sequentially. If the first one returns an Error response, the second one is never executed.