Module oak_client::proto

Modules

request_wrapper

Nested message and enum types in RequestWrapper.

response_wrapper

Nested message and enum types in ResponseWrapper.

streaming_session_client

Generated client implementations.

streaming_session_server

Generated server implementations.

Structs

ApplicationData

AttestationEvidence

AttestationEvidence contains the information that the untrusted launcher provides to the client in response to its request for the enclave’s public key(s). The name is chosen to match the RATS terminology (see https://www.rfc-editor.org/rfc/rfc9334.html#name-evidence).

BinaryAttestation

GetPublicKeyRequest

GetPublicKeyResponse

InvokeRequest

InvokeResponse

RequestWrapper

ResponseWrapper