Module oak_client::proto::oak::attestation::v1

Modules

attestation_results

Nested message and enum types in AttestationResults.

binary_reference_value

Nested message and enum types in BinaryReferenceValue.

endorsements

Nested message and enum types in Endorsements.

expected_digests

Nested message and enum types in ExpectedDigests.

expected_values

Nested message and enum types in ExpectedValues.

extracted_evidence

Nested message and enum types in ExtractedEvidence.

kernel_binary_reference_value

Nested message and enum types in KernelBinaryReferenceValue.

reference_values

Nested message and enum types in ReferenceValues.

regex_reference_value

Nested message and enum types in RegexReferenceValue.

root_layer_data

Nested message and enum types in RootLayerData.

text_expected_value

Nested message and enum types in TextExpectedValue.

text_reference_value

Nested message and enum types in TextReferenceValue.

Structs

AmdAttestationReport

Values extracted from an AMD SEV-SNP attestation report.

AmdSevExpectedValues

AmdSevReferenceValues

ApplicationKeys

Keys used by the application to derive encryption session keys and to sign arbitrary data. Each of the certificates contains the final layer’s measurement as additional claims.

ApplicationLayerData

Values extracted from the evidence that represents an application running under the Oak Restricted Kernel.

ApplicationLayerEndorsements

ApplicationLayerExpectedValues

The expected binary and configuration digests for an application layer.

ApplicationLayerReferenceValues

Represents an application running under Oak Restricted Kernel.

AttestationResults

Represents a verification result. Can be extended to return certain measurements and other detail to the client for further processing. Nomenclature follows RFC 9334.

BinaryReferenceValue

CbApplicationLayerExpectedValues

Represents digest of application task config.

CbApplicationLayerReferenceValues

Represents digest of application task config.

CbData

Reserved for future use.

CbEndorsements

CbExpectedValues

CbReferenceValues

CbSystemLayerExpectedValues

Represents digest of application task config.

CbSystemLayerReferenceValues

CertificateAuthority

Message for passing embedded certificate authority information between layers. Will never appear in the evidence that is sent to the client.

ContainerLayerData

Values extracted from the evidence that represents the Container Runtime Bundle used in Oak Containers.

ContainerLayerEndorsements

ContainerLayerExpectedValues

The expected bundle and configuration digests for a container layer.

ContainerLayerReferenceValues

DiceData

Message that is sent between DICE layers.

Digests

Verifies that the field contains at least one of the given digests. No checks are performed if this is empty. A match in at least one digest is considered a success.

EndorsementReferenceValue

Verifies the transparency log entry, including signatures and the digest.

Endorsements

Endorsement contains statements that some entity (e.g. a hardware provider) vouches for the integrity of claims about the TEE or the software running on it.

Event

Represents an event intended for inclusion in attestation. For example, in an attested measured boot, each event is a reference to the code identity of the boot layer being launched next. An Event message contain what’s necessary for an attestation verifier to verify the Event against a Reference Value. TODO: b/333748757 - Make other CB layers use this definition.

EventLog

A sequence of Events intended for inclusion in attestation evidence.

Evidence

Attestation Evidence used by the client to the identity of firmware and software running inside a Trusted Execution Environment.

ExpectedDigests

ExpectedRegex

ExpectedStringLiterals

ExpectedValues

ExtractedEvidence

Evidence values extracted from attestation evidence during verification.

FakeAttestationReport

Values extracted from a fake attestation report when not running in a TEE.

FileReferenceValue

Reference value for a file including its digests.

FirmwareAttachment

The Transparent Release attachment for Oak Stage 0. Measurements are produced with https://github.com/project-oak/oak/snp_measurement.

InsecureExpectedValues

InsecureReferenceValues

IntelTdxAttestationReport

Values extracted from an Intel TDX attestation report.

IntelTdxExpectedValues

IntelTdxReferenceValues

KernelAttachment

The Transparent Release attachment for Oak Containers Linux kernel. Measurements are produced with https://github.com/project-oak/oak/oak_kernel_measurement.

KernelBinaryReferenceValue

Follows the lines of BinaryReferenceValue, but provides a custom proto to facilitate the digest matching.

KernelDigests

Similar to the Digests message, but allows to specify digests for the split components of the bzImage separately.

KernelExpectedValues

The expected values for kernel image and setup data, computed from previously provided endorsements and reference values.

KernelLayerData

Values extracted from the the kernel layer evidence, as measured by stage0.

KernelLayerEndorsements

KernelLayerExpectedValues

Reference values of the kernel layer, as measured by stage0.

KernelLayerReferenceValues

Reference values of the kernel layer, as measured by stage0.

LayerEvidence

DICE layer evidence containing a certificate signed by the previous layer.

OakContainersData

Values extracted from the evidence for an Oak Containers instance.

OakContainersEndorsements

OakContainersExpectedValues

OakContainersReferenceValues

OakRestrictedKernelData

Values extracted from the evidence for a restricted kernel application.

OakRestrictedKernelEndorsements

OakRestrictedKernelExpectedValues

OakRestrictedKernelReferenceValues

RawDigests

ReferenceValues

Regex

RegexReferenceValue

RootLayerData

Values extracted from the root layer evidence.

RootLayerEndorsements

RootLayerEvidence

Evidence generated by the Layer0.

RootLayerExpectedValues

RootLayerReferenceValues

SkipVerification

Stage0Measurements

All the related measurements for Stage 0.

StringLiterals

A match in at least one value is considered a success. At least one value must be specified, otherwise verification fails.

StringReferenceValue

Verifies that a particular string is equal to at least one of the specified ones. No checks are performed if this is empty.

SystemLayerData

Values extracted from the evidence that represents the Oak Containers system image.

SystemLayerEndorsements

SystemLayerExpectedValues

The expected binary digests for a system layer image.

SystemLayerReferenceValues

TcbVersion

The versions of the components in the AMD SEV-SNP platform Trusted Compute Base (TCB).

TextExpectedValue

TextReferenceValue

Reference value to match text via endorsement, or directly via constants or a regular expression.

TransparentReleaseEndorsement

Endorsement for binaries from the Transparent Release process. <https://github.com/project-oak/oak/blob/main/docs/release.md>

VerificationSkipped

Enums

TeePlatform

This proto defines the layered DICE Attestation Evidence.