Module oak_client::proto::oak::crypto::v1

Structs

AeadEncryptedMessage

Message encrypted with Authenticated Encryption with Associated Data (AEAD). <https://datatracker.ietf.org/doc/html/rfc5116>

EncryptedRequest

Request message encrypted using Hybrid Public Key Encryption (HPKE). <https://www.rfc-editor.org/rfc/rfc9180.html>

EncryptedResponse

Response message encrypted Hybrid Public Key Encryption (HPKE), which uses a response key generated as part of bidirectional encryption. <https://www.rfc-editor.org/rfc/rfc9180.html#name-bidirectional-encryption>

SessionKeys

Envelope containing session keys required to encrypt/decrypt messages within a secure session. Needed to serialize contexts in order to send them over an RPC.

Signature