Module oak_client::proto::oak::session::v1

Modules

request_wrapper

Nested message and enum types in RequestWrapper.

response_wrapper

Nested message and enum types in ResponseWrapper.

streaming_session_client

Generated client implementations.

Structs

ApplicationData

AttestationBundle

AttestationBundle contains the information that the untrusted launcher provides to the client in response to its request for the enclave’s public key(s).

AttestationEndorsement

AttestationEndorsement contains statements that some entity (e.g., a hardware provider) vouches for the integrity of claims about the TEE or the software running on it. (see https://www.rfc-editor.org/rfc/rfc9334.html#name-endorsements).

AttestationEvidence

AttestationEvidence contains all the information that originates in the TEE and can serve as claims about the Target Environment. The name is chosen to match the RATS terminology (see https://www.rfc-editor.org/rfc/rfc9334.html#name-evidence).

GetEncryptionKeyRequest

GetEncryptionKeyResponse

GetPublicKeyRequest

GetPublicKeyResponse

InvokeRequest

InvokeResponse

RequestWrapper

ResponseWrapper