Struct oak_client::transport::GrpcStreamingTransport

pub struct GrpcStreamingTransport { /* private fields */ }

A Transport implementation that uses a single gRPC streaming session to get the evidence and then invokve the desired request.

Implementations

impl GrpcStreamingTransport

pub async fn new<Fut>( stream_creator: impl FnOnce(Receiver<RequestWrapper>) -> Fut ) -> Result<Self>
where Fut: Future<Output = Result<Response<Streaming<ResponseWrapper>>>>,

Create a new GrpcStreamingTransport.

The provided stream_creator will be immediately invokved to create a new stream session, and all actions on the newly created instance will use that stream. If the stream dies, a new client/transport pair will need to be created.

The stream_creator is a closure to create a new stream. Typically this will be used with a gRPC client instance that exposes a bi-directional streaming method.

For example, if you have a gRPC service like:

service {
    rpc MyMethod(stream RequestWrapper) returns (stream ResponseWrapper);
}

You will call:

let transport =
  GrpcStreamingTransport::new(|rx| client.my_method(rx))