Module oak_client::interceptors

Modules

auth

label

Structs

CombinedInterceptor
This struct is created by the combine method. See its documentation for more.

Traits

Interceptor
Trait for interceptors, to facilitate combining until https://github.com/hyperium/tonic/issues/499 offers a better solution.

Functions

combine
Combines the two provided interceptor, executing the first one, and if that succeeds then also the second one, sequentially. If the first one returns an Error response, the second one is never executed.